control_container/requirements.txt
flask
pyshark
mysql-connector-python